Pass down data as props to dashboard on profile page
